// src/types/types.ts
export type MenuData = {
    path: string;
    name: string;
    img: string;
}

export type Product = {
    id: string | number;
    posImg: string;
    posTitle: string;
    posQuality: number | string;
    memName: string;
    memPhone: string;
    memStatus: boolean;
    memTime: string;
    memType: string;
    visitDate: string;
    visitName: string;
    visitTime: string;
    pamentTerm: string;
    paymentCheck: boolean;
    paymentDate: string;
    paymentPaid: string;
    paymentTotal: string;
}